+++ GooglePay/src/main/java/com/braintreepayments/api/ReadyForGooglePayRequest.java
@@ -1,7 +1,5 @@
 package com.braintreepayments.api;
 
-import androidx.fragment.app.FragmentActivity;
-
 /**
  * Optional parameters to use when checking whether Google Pay is supported and set up on the customer's device.
  */
@@ -10,7 +8,7 @@ public class ReadyForGooglePayRequest {
     private boolean existingPaymentMethodRequired;
 
     /**
-     * If set to true, then the {@link GooglePayClient#isReadyToPay(FragmentActivity, ReadyForGooglePayRequest, GooglePayIsReadyToPayCallback)}
+     * If set to true, then the {@link GooglePayClient#isReadyToPay(android.content.Context, ReadyForGooglePayRequest, GooglePayIsReadyToPayCallback)}
      * method will call the listener with true if the customer is ready to pay with one or more of your
      * supported card networks.
      *